桌面/app/requirements.txt
fastapi
uvicorn[standard]
httpx
websockets
pillow
json
logging
